+++ upload-api/src/Repository/MaterialRepository.php
@@ -22,15 +22,22 @@ class MaterialRepository extends ServiceEntityRepository
      *
      * @param int $agencyId
      *   Agency ID
+     * @param bool $isNotUploaded
+     *   Filter base on those that do not have covers marked as uploaded
      *
      * @return Query
      *   The query build
      */
-    public function getByAgencyId(int $agencyId): Query
+    public function getByAgencyId(int $agencyId, bool $isNotUploaded): Query
     {
         $queryBuilder = $this->createQueryBuilder('m')
             ->where('m.agencyId = :agencyId')
             ->setParameter(':agencyId', $agencyId);
+
+        if ($isNotUploaded) {
+            $queryBuilder->join('m.cover', 'c')
+                ->where('c.isUploaded = 0');
+        }
 
         return $queryBuilder->getQuery();
     }
@@ -38,12 +45,20 @@ class MaterialRepository extends ServiceEntityRepository
     /**
      * Get all materials.
      *
+     * @param bool $isNotUploaded
+     *   Filter base on those that do not have covers marked as uploaded
+     *
      * @return Query
      *   The query build
      */
-    public function getAll(): Query
+    public function getAll(bool $isNotUploaded): Query
     {
         $queryBuilder = $this->createQueryBuilder('m');
+
+        if ($isNotUploaded) {
+            $queryBuilder->join('m.cover', 'c')
+                ->where('c.isUploaded = 0');
+        }
 
         return $queryBuilder->getQuery();
     }