+++ upload-api/CHANGELOG.md
@@ -11,10 +11,16 @@ See [keep a changelog](https://keepachangelog.com/en/1.0.0/) for information abo
 
 ## [Unreleased]
 
+## [2.2.9] - 2022-06-13
+
+### Changed
+- Updated minor version of Symfony.
+- Added support for filtering on isUploaded on re-queue command.
+
 ## [2.2.8] - 2022-06-10
 
 ### Changed
-- Updated app:database:cleanup command to mark uploaded correctly. 
+- Updated app:database:cleanup command to mark uploaded correctly.
 
 ## [2.2.7] - 2022-06-09